+++ lib/pulsar/components/field.ex
@@ -542,6 +542,7 @@ defmodule Pulsar.Components.Field do
       id={@field_id}
       name={@field_name}
       value={@field_value}
+      form={@form}
       variant={@variant}
       color={@color}
       size={@size}
@@ -567,6 +568,7 @@ defmodule Pulsar.Components.Field do
       filter={@filter}
       async={@async}
       display={@display}
+      form={@form}
       labelled_externally
       variant={@variant}
       color={@color}
@@ -587,6 +589,7 @@ defmodule Pulsar.Components.Field do
       field={@field}
       id={@field_id}
       mode="single"
+      form={@form}
       labelled_externally
       months={@months}
       min={@min}
@@ -619,6 +622,7 @@ defmodule Pulsar.Components.Field do
       end_field={@end_field}
       id={@field_id}
       mode="range"
+      form={@form}
       labelled_externally
       months={@months}
       min={@min}